refactor(util): 重构 VersionUtil 并添加新功能
- 重命名 VersionBeanUtil 为 VersionUtil
- 移动 resolveLocalQQ 和 resolveLocalTIM 函数到 VersionUtil
- 添加 findActivity 扩展函数到 InfoUtil
- 更新 MainActivity 中的错误处理和版本解析逻辑
- 新增疑似撤包提示字符串到多个语言文件
