+++ app/src/main/java/com/xiaoniu/qqversionlist/util/InfoUtil.kt
@@ -21,6 +21,7 @@ package com.xiaoniu.qqversionlist.util
 import android.app.Activity
 import android.app.AlertDialog
 import android.content.Context
+import android.content.ContextWrapper
 import android.content.Intent
 import android.os.Build
 import android.provider.Settings
@@ -66,12 +67,13 @@ object InfoUtil {
      * @param isCustomMessage 布尔值，表示是否使用自定义的错误消息（不在前台显示跟踪堆栈），默认为 false
      * @param isShowSystemNotifSetting 布尔值，表示是否显示前往系统通知设置按钮，默认为 false
      */
-    fun Activity.dialogError(
+    fun Context.dialogError(
         e: Exception,
         isCustomMessage: Boolean = false,
         isShowSystemNotifSetting: Boolean = false
     ) {
-        runOnUiThread {
+        val activity = findActivity()
+        activity?.runOnUiThread {
             val message = if (isCustomMessage) e.message else buildString {
                 appendLine("如需反馈，请前往 GitHub 仓库报告 Issue(s) 并随附以下信息：\n")
                 appendLine(e.stackTraceToString())
@@ -138,5 +140,24 @@ object InfoUtil {
                 setSpan(URLSpan(url), start, end, SpannableString.SPAN_EXCLUSIVE_EXCLUSIVE)
             }
         }
+    }
+
+    /**
+     * 在 `Context` 中查找 `Activity`。
+     *
+     * 该函数通过不断地获取 `baseContext` 来遍历 `ContextWrapper` 链，直到找到一个 `Activity` 或无法继续遍历。
+     * 这种方法用于获取当前上下文相关的 `Activity` 实例，以便执行某些操作。
+     *
+     * @return 如果找到 `Activity` 则返回该 `Activity` 实例，否则返回null。
+     */
+    fun Context.findActivity(): Activity? {
+        var context = this
+        while (context is ContextWrapper) {
+            if (context is Activity) {
+                return context
+            }
+            context = context.baseContext
+        }
+        return null
     }
 }
