+++ app/src/main/java/com/xiaoniu/qqversionlist/util/StringUtil.kt
@@ -19,6 +19,7 @@
 package com.xiaoniu.qqversionlist.util
 
 import android.app.Activity
+import android.content.Context
 import android.content.pm.PackageInfo
 import com.google.gson.Gson
 import com.xiaoniu.qqversionlist.util.FileUtil.ZipFileCompat
@@ -143,10 +144,10 @@ object StringUtil {
      * 从给定的 `PackageInfo` 对象中获取 `qua.ini` 文件的内容
      *
      * @param packageInfo 包含应用信息的 `PackageInfo` 对象，用于访问应用的资源
-     * @param activity 用于显示错误对话框的 `Activity` 对象
+     * @param Context 用于显示错误对话框的 `Context` 对象
      * @return 返回 `qua.ini` 文件的内容作为字符串，如果发生任何错误或文件不存在则返回null
      */
-    fun Activity.getQua(packageInfo: PackageInfo): String? {
+    fun Context.getQua(packageInfo: PackageInfo): String? {
         val sourceDir = packageInfo.applicationInfo?.sourceDir ?: return null
         val file = File(sourceDir)
         if (!file.exists()) return null
