+++ app/src/main/res/values-en-rUS/strings.xml
@@ -189,4 +189,5 @@
     <string name="successInGetting">Success in Getting</string>
     <string name="getWeixinAlphaConfig404">It seems that the Tencent server configuration file was not accessed successfully, possibly because there is no current alpha version of Weixin or the beta version has been removed.</string>
     <string name="getWeixinAlphaConfigLink404">It seems that the download address was not successfully accessed, possibly because the current Weixin alpha version has been removed.</string>
+    <string name="suspectedPackageWithdrawal">Suspected Package Withdrawal</string>
 </resources>
