+++ app/src/main/res/values-ru-rRU/strings.xml
@@ -189,4 +189,5 @@
     <string name="successInGetting">获取成功</string>
     <string name="getWeixinAlphaConfig404">似乎未成功访问腾讯服务器配置文件，可能是微信当前没有测试版或测试版已撤包。</string>
     <string name="getWeixinAlphaConfigLink404">似乎未成功访问此下载地址，可能是微信当前测试版已撤包。</string>
+    <string name="suspectedPackageWithdrawal">疑似撤包</string>
 </resources>
